+++ data_management/lib/mysql_connection.py
@@ -2,10 +2,13 @@ import os
 import mysql.connector
 from dotenv import load_dotenv
 import logging
+import requests
 
+load_dotenv()
+slack_passcode = os.environ.get('slack_passcode')
 logger = logging.getLogger("lib-MYSQLConnection")
 logging.basicConfig(level=logging.ERROR)
-
+slack_url = "https://hooks.slack.com/services/" + slack_passcode
 
 class MYSQLConnection:
     def __init__(self):
@@ -18,12 +21,26 @@ class MYSQLConnection:
         self.password = None
         self.database_name = None
 
+        self.tableau_host = None
+        self.tableau_port = None
+        self.tableau_user = None
+        self.tableau_password = None
+        self.tableau_database_name = None
+
         try:
             self.host = os.environ["mysql_host"]
             self.port = os.environ["mysql_port"]
             self.user = os.environ["mysql_user"]
             self.password = os.environ["mysql_pwd"]
             self.database_name = os.environ["mysql_db"]
+
+            self.tableau_host = os.environ["tableau_host"]
+            self.tableau_port = os.environ["tableau_port"]
+            self.tableau_user = os.environ["tableau_user"]
+            self.tableau_password = os.environ["tableau_password"]
+            self.tableau_database_name = os.environ["tableau_database_name"]
+            self.tableau_ca_file = os.environ["tableau_ca_file"]
+
         except:
             logger.warning(
                 "Can't load environment variables from docker... trying local .env file instead..."
@@ -38,6 +55,13 @@ class MYSQLConnection:
             self.user = os.environ.get("mysql_user")
             self.password = os.environ.get("mysql_pwd")
             self.database_name = os.environ.get("mysql_db")
+
+            self.tableau_host = os.environ.get("tableau_host")
+            self.tableau_port = os.environ.get("tableau_port")
+            self.tableau_user = os.environ.get("tableau_user")
+            self.tableau_password = os.environ.get("tableau_password")
+            self.tableau_database_name = os.environ.get("tableau_database_name")
+            self.tableau_ca_file =  os.environ.get("tableau_ca_file")
         except:
             logger.warning("Can't load environment variables from local .env file")
 
@@ -57,11 +81,28 @@ class MYSQLConnection:
                 port=self.port,
                 password=self.password,
                 database=self.database_name,
+                autocommit=True
             )
             self.database.get_warnings = True
             return self.database
         except Exception as error:
             logger.error("Can't connect to MySQL: ", exec_info=error)
+            os.sys.exit()
+
+    def get_tableau_db_connection(self):
+        try:
+            self.database = mysql.connector.connect(
+                host=self.tableau_host,
+                user=self.tableau_user,
+                port=self.tableau_port,
+                password=self.tableau_password,
+                database=self.tableau_database_name,
+                ssl_ca=self.tableau_ca_file
+            )
+            self.database.get_warnings = True
+            return self.database
+        except Exception as error:
+            logger.error("Can't connect to MySQL: ", exc_info=error)
             os.sys.exit()
 
     def insert_data(self, sql, data):
@@ -72,7 +113,13 @@ class MYSQLConnection:
             if warning is not None:
                 print(warning)
         except:
-            logger.error(f"Cannot insert with query: {sql}; and the data: {data}")
+            message = f"Error: Cannot insert with query: {sql}; and the data: {data}"
+            logger.error(message)
+            requests.post(
+                slack_url,
+                headers={'Content-type': 'application/json', },
+                data='{"text":"' + message + '"}'
+            )
         finally:
             self.database.commit()
             self.cursor.close()
@@ -86,7 +133,13 @@ class MYSQLConnection:
                 data.append(row)
             return data
         except:
-            logger.error("Can't get data_management data.")
+            message = "Error: Can't get data_management data."
+            logger.error(message)
+            requests.post(
+                slack_url,
+                headers={'Content-type': 'application/json', },
+                data='{"text":"' + message + '"}'
+            )
         finally:
             self.cursor.close()