+++ data_management/services/dlu_management.py
@@ -1,3 +1,5 @@
+from datetime import datetime
+
 from lib.mysql_connection import MYSQLConnection
 from lib.mongo_connection import MongoConnection
 import logging
@@ -55,7 +57,7 @@ class DluManagement:
             )
             self.db.insert_data(
                 "INSERT INTO redcap_participant ( "
-                + "redcap_id, redcap_np_gender, redcap_age_binned, redcap_tissue_type, "
+                + "redcap_id, redcap_np_gender, redcap_age_binned, redcap_enrollment_category, "
                 + "redcap_protocol, redcap_sample_type, redcap_tissue_source, redcap_clinical_data,"
                   "redcap_exp_aki_kdigo, redcap_exp_race, redcap_exp_alb_cat_most_recent,"
                   "redcap_mh_ht_yn, redcap_mh_diabetes_yn, redcap_exp_has_med_raas, "
@@ -64,10 +66,12 @@ class DluManagement:
                 + "VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)",
                 (list(redcap_participant.values())),
             )
+            return 1
         else:
             logger.warning(
                 f"redcap participant with id: {redcap_participant['redcap_id']} already exists, skipping insert"
             )
+            return 0
 
     def get_participant_by_redcap_id(self, redcap_id: str):
         return self.db.get_data(
@@ -78,8 +82,8 @@ class DluManagement:
     def insert_dlu_package(self, dpi_values: tuple, dmd_values: tuple):
         logger.info(f"inserting DLU package with id: {dpi_values[0]}")
         query1 = ("INSERT INTO dlu_package_inventory (dlu_package_id, dlu_created, dlu_submitter, dlu_tis, "
-                 + "dlu_packageType, dlu_subject_id, dlu_error, dlu_lfu, globus_dlu_status) "
-                 + "VALUES(%s, %s, %s, %s, %s, %s, %s, %s, %s)")
+                 + "dlu_packageType, dlu_subject_id, dlu_error, dlu_lfu, dlu_upload_type, globus_dlu_status) "
+                 + "VALUES(%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)")
         self.db.insert_data(query1, dpi_values)
         dmd_values = ((dmd_values[0],) + dmd_values)
         query2 = ("INSERT INTO dmd_data_manager (id, dlu_package_id, redcap_id, known_specimen, "
@@ -99,15 +103,31 @@ class DluManagement:
             self.db.insert_data(query, values)
 
     def insert_dlu_file(self, values):
-        query = "INSERT INTO dlu_file (dlu_fileName, dlu_package_id, dlu_file_id, dlu_filesize, dlu_md5checksum, dlu_metadata) VALUES(%s, %s, %s, %s, %s, %s)"
+        query = "INSERT INTO dlu_file (dlu_fileName, dlu_package_id, dlu_file_id, dlu_filesize, dlu_md5checksum, dlu_modified_at, dlu_metadata) VALUES(%s, %s, %s, %s, %s, %s, %s)"
         self.db.insert_data(query, values)
         return query % values
 
-    def insert_dlu_files(self, package_id: str, file_list: List[DLUFile]):
+    def insert_dlu_files(self, package_id: str, file_list: List[DLUFile]) -> dict:
         logger.info(f"Inserting files for package {package_id}")
+        existing_files = self.get_files_by_package_id(package_id)
+        unmodified_files = []
+        if existing_files is not None and len(existing_files) > 0:
+            logger.info(f"Deleting existing files for package {package_id}")
+            self.delete_files_by_package_id(package_id)
         for file in file_list:
-            query_string = self.insert_dlu_file((file.name, package_id, file.file_id, file.size, file.checksum, json.dumps(file.metadata)))
+            for existing_file in existing_files:
+                if existing_file["dlu_fileName"] == file.name:
+                    file.modified_at = existing_file["dlu_modified_at"]
+                    if file.checksum != existing_file["dlu_md5checksum"]:
+                        file.modified_at = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+                    else:
+                        unmodified_files.append(file)
+                    file.file_id = existing_file["dlu_file_id"]
+                    existing_files.remove(existing_file)
+            query_string = self.insert_dlu_file((file.name, package_id, file.file_id, file.size, file.checksum, file.modified_at, json.dumps(file.metadata)))
             logger.info(query_string)
+
+        return {"files": file_list, "deleted_files": existing_files, "unmodified_files": unmodified_files}
 
     def get_ready_to_move(self, package_id: str):
         package_record = self.db.get_data(
@@ -160,6 +180,53 @@ class DluManagement:
         else:
             return None
 
+    def get_biopsy_tracking(self):
+        result = self.db.get_data(
+            "select * from biopsy_tracking_v"
+        )
+        return result
+
+    def get_data_manager_data(self):
+        result = self.db.get_data(
+            """
+                select dm.id, dm.dlu_package_id, dm.dlu_created, dm.dlu_submitter, dm.dlu_tis, dm.dlu_packageType, dm.dlu_subject_id, dm.dlu_error, dm.redcap_id, dm.known_specimen, dm.user_package_ready, dm.package_validated, dm.ready_to_move_from_globus, dm.globus_dlu_status, dm.package_status, dm.current_owner, dm.ar_promotion_status, dm.sv_promotion_status, dm.release_version, r.release_date, dm.removed_from_globus, dm.notes 
+                from data_manager_data_v dm
+                left outer join `release` r on dm.release_version = r.release_version 
+            """
+        )
+        return result
+
+    def get_package(self, package_id: str) -> dict:
+        result = self.db.get_data("SELECT * dlu_package_inventory WHERE dlu_package_id = %s", (package_id,))
+        if result:
+            return result[0]
+        else:
+            return None
+
+    def get_files_by_package_id(self, package_id: str):
+        return self.db.get_data("SELECT * FROM dlu_file WHERE dlu_package_id = %s", (package_id,))
+
+    def delete_files_by_package_id(self, package_id: str):
+        return self.db.get_data("DELETE FROM dlu_file WHERE dlu_package_id = %s", (package_id,))
+
+    def get_equal_num_rows(self):
+        result = self.db.get_data("SELECT (SELECT COUNT(*) FROM slide_manifest_import) = (SELECT COUNT(*) FROM slide_scan_curation) AS equal_num_rows")
+        return result[0]["equal_num_rows"]
+    
+    def get_new_slide_manifest_import_rows(self):
+        return self.db.get_data("SELECT * FROM slide_manifest_import WHERE image_id NOT IN (SELECT image_id FROM slide_scan_curation)")
+
+    def get_spectrack_redcap_record_id(self, kit_id):
+        result = self.db.get_data("SELECT spectrack_redcap_record_id FROM spectrack_specimen WHERE spectrack_specimen_kit_id = %s LIMIT 1", (kit_id,))
+        if len(result) > 0 and "spectrack_redcap_record_id" in result[0]:
+            return result[0]["spectrack_redcap_record_id"]
+        else:
+            return None
+
+    def insert_into_slide_scan_curation(self, values):
+        query = "INSERT INTO slide_scan_curation (image_id, kit_id, redcap_id) VALUES (%s, %s, %s)"
+        self.db.insert_data(query, values)
+        return query % values
 
 if __name__ == "__main__":
     dlu_management = DluManagement()