+++ data_management/services/dlu_package_inventory.py
@@ -19,7 +19,7 @@ class DLUPackageInventory:
         
     def get_dlu_file(self, status):
         return self.db.get_data(
-            "SELECT * FROM data_management.data_manager_data_v WHERE ready_to_move_from_globus = %s AND globus_dlu_status IS NULL",
+            'SELECT * FROM data_management.data_manager_data_v WHERE ready_to_move_from_globus = %s AND (globus_dlu_status IS NULL OR globus_dlu_status = "recalled")',
             (status,)
         )
     
@@ -27,6 +27,11 @@ class DLUPackageInventory:
         return self.db.insert_data(
             'UPDATE data_management.data_manager_data_v SET globus_dlu_status = "waiting" WHERE ready_to_move_from_globus = %s AND dlu_package_id = %s',
             (status, package_id,)
+        )
+        
+    def get_waiting_files(self):
+        return self.db.get_data(
+            'Select * from data_management.data_manager_data_v where globus_dlu_status = "waiting" and ready_to_move_from_globus = "yes"'
         )
     
     def get_ready_packages(self):