+++ data_management/services/dlu_utils.py
@@ -15,7 +15,8 @@ def dlu_package_dict_to_dpi_tuple(dlu_inventory: dict):
         dlu_inventory["dluSubjectId"],
         dlu_inventory["dluError"],
         dlu_inventory["dluLfu"],
-        dlu_inventory["globusDluStatus"]
+        dlu_inventory['dluUploadType'],
+        None
     )
 
 def dlu_package_dict_to_dmd_tuple(dlu_inventory: dict):