+++ data_management/watch_files.py
@@ -38,6 +38,22 @@ class DLUWatcher:
         else:
             self.update_files_for_globus(files)
             self.move_packages_to_DLU(files)
+    
+    def watch_for_side_manifest_records(self):
+        equal_num_rows = self.dlu_management.get_equal_num_rows()
+        if equal_num_rows == 1:
+            logger.info("No new records found in slide_manifest_import")
+        else:
+            self.update_slide_scan_curation()
+    
+    def update_slide_scan_curation(self):
+        logger.info("Importing new row(s) into slide_scan_curation")
+        new_records = self.dlu_management.get_new_slide_manifest_import_rows()
+        for record in new_records:
+            redcap_id = self.dlu_management.get_spectrack_redcap_record_id(record["outside_acc"])
+            slide_scan_tuple = (record["image_id"], record["outside_acc"], redcap_id)
+            query_string = self.dlu_management.insert_into_slide_scan_curation(slide_scan_tuple)
+            logger.info(query_string)
             
     def update_files_for_globus(self, files):
         for index, file_result in enumerate(files):
@@ -50,6 +66,15 @@ class DLUWatcher:
             file.path = self.dlu_file_handler.split_path(file.path)['file_path']
             dlu_files.append(file)
         return dlu_files
+    
+    def pickup_waiting_files(self):
+        files_in_waiting = self.db.get_waiting_files()
+        if len(files_in_waiting) == 0:
+            return  logger.info(
+                "No records were found with status 'waiting'"
+            )
+        else:
+            self.move_packages_to_DLU(files_in_waiting)
 
     def move_packages_to_DLU(self, packages):
         file_list = None
@@ -79,20 +104,23 @@ class DLUWatcher:
                 file_list = self.dlu_file_handler.match_files(top_level_subdir)
             else:
               file_list = self.dlu_file_handler.match_files(package_id)
-                            
+
             self.dlu_file_handler.copy_files(package_id, self.process_file_paths(directory_info.file_details))
-            self.dlu_file_handler.chown_dir(package_id, file_list)
-            self.dlu_management.insert_dlu_files(package_id, file_list)
+            self.dlu_file_handler.chown_dir(package_id, file_list, int(os.environ['dlu_user']))
+            file_info = self.dlu_management.insert_dlu_files(package_id, file_list)
             self.dlu_management.update_dlu_package(package_id, { "globus_dlu_status": "success" })
             self.dlu_management.update_dlu_package(package_id, { "ready_to_move_from_globus": "done" })
-            self.dlu_mongo.update_package_files(package_id, file_list)
+            self.dlu_mongo.update_package_files(package_id, file_info)
+            
             self.dlu_state.set_package_state(package_id, PackageState.UPLOAD_SUCCEEDED)
             self.dlu_state.clear_cache()
 
 
 
 if __name__ == "__main__":
-    while True:
-        dlu_watcher = DLUWatcher()
+    dlu_watcher = DLUWatcher()
+    dlu_watcher.pickup_waiting_files()
+    while True:    
         dlu_watcher.watch_for_files()
-        time.sleep(60)
+        time.sleep(60) 
+        dlu_watcher.watch_for_side_manifest_records()