+++ changelog.md
@@ -1,9 +1,19 @@
 # Changelog
 
 
+## Release 1.9 [unreleased]
+
+- Load Tableau database with biopsy_tracker data
+- Insert dlu_upload_type to dlu_package_inventory table
+- Create recalled packages endpoint
+- Update Multi Modal package name
+- Tweak bulk upload fields
+
+
 ## Release 1.8.1 [Released 11/8/2024]
 Brief summary of what's in this release:
 Fix to the dlu-watcher script to generate a new UUID for each file, rather than one at start up. This caused issues with moving files
+
 
 ### Breaking changes